+++ Classes/Templating/TemplateHelper.php
@@ -463,38 +463,6 @@ class TemplateHelper extends SalutationSwitcher
     }
 
     /**
-     * Sets a cached configuration value that will be used when a new instance is created.
-     *
-     * This function is intended to be used for testing purposes only.
-     *
-     * @param string $key key of the configuration property to set, must not be empty
-     * @param mixed $value value of the configuration property, may be empty or zero
-     *
-     * @deprecated will be removed in oelib 4.0
-     */
-    public static function setCachedConfigurationValue(string $key, $value): void
-    {
-        $pageUid = PageFinder::getInstance()->getPageUid();
-        if (!isset(self::$cachedConfigurations[$pageUid])) {
-            self::$cachedConfigurations[$pageUid] = [];
-        }
-
-        self::$cachedConfigurations[$pageUid][$key] = $value;
-    }
-
-    /**
-     * Purges all cached configuration values.
-     *
-     * This function is intended to be used for testing purposes only.
-     *
-     * @deprecated will be removed in oelib 4.0
-     */
-    public static function purgeCachedConfigurations(): void
-    {
-        self::$cachedConfigurations = [];
-    }
-
-    /**
      * Gets the configuration.
      *
      * @return array configuration array, might be empty